refactor (Triggers): TriggersStore and TriggersAPI renamed and moved where StorageAPI is
